Add files via upload
The start of my Uno project in Java. Includes a working player class where they are able to be given a random hand of 7 cards, place cards that match with the top card of the discard pile, and change the colour after placing a wild card. Still need to finish CPU control and the option to stack multiple cards of the same number in one hand.